promoters part1 tests

+++ database/migrations/2023_04_18_203634_add_partner_id_field_to_piers_table.php
@@ -27,10 +27,12 @@ class AddPartnerIdFieldToPiersTable extends Migration
      */
     public function down()
     {
-        Schema::table('piers', function (Blueprint $table) {
-            $table->dropColumn('external_id');
-            $table->dropColumn('external_parent_id');
-            $table->dropColumn('source');
-        });
+        if (Schema::hasColumns('piers', ['external_id', 'external_parent_id', 'source'])) {
+            Schema::table('piers', function (Blueprint $table) {
+                $table->dropColumn('external_id');
+                $table->dropColumn('external_parent_id');
+                $table->dropColumn('source');
+            });
+        }
     }
 }
